1. Unify two pipelines into one. 2. Make sure the configs can be properly overwritten 3. Add tests for the example in GPT2, which only tests the pipeline construction via YAML (GPT is too slow to be tested.)
