+++ examples/gpt2/sample_multipack_pipeline_gpt.yml
@@ -9,27 +9,25 @@ Reader:
 
 Processors:
   - type: "forte.processors.TextGenerationProcessor"
-    hparams:
-      overwrite_configs:
-        max_decoding_length: 128
-        temperature: 0.7
-        top_k: 40
-        pretrained_model_name: "gpt2-small"
-        input_pack_name: "input_src"
-        output_pack_name: "output_tgt"
+    configs:
+      max_decoding_length: 128
+      temperature: 0.7
+      top_k: 40
+      pretrained_model_name: "gpt2-small"
+      input_pack_name: "input_src"
+      output_pack_name: "output_tgt"
+      batcher:
         batch_size: 64
   - type: "forte.processors.nltk_processors.NLTKWordTokenizer"
-    hparams:
-      overwrite_configs:
-        selector:
-          type: "forte.data.selector.NameMatchSelector"
-          kwargs:
-            select_name: "output_tgt"
+    configs:
+      selector:
+        type: "forte.data.selector.NameMatchSelector"
+        kwargs:
+          select_name: "output_tgt"
   - type: "forte.processors.nltk_processors.NLTKPOSTagger"
-    hparams:
-      overwrite_configs:
-        selector:
-          type: "forte.data.selector.NameMatchSelector"
-          kwargs:
-            select_name: "output_tgt"
+    configs:
+      selector:
+        type: "forte.data.selector.NameMatchSelector"
+        kwargs:
+          select_name: "output_tgt"
