+++ examples/pipelines/stanford_nlp_pipeline.py
@@ -17,6 +17,7 @@ import os
 from termcolor import colored
 from texar.torch import HParams
 
+from forte.data.data_pack import DataPack
 from forte.data.readers import StringReader
 from forte.pipeline import Pipeline
 from forte.processors.stanfordnlp_processor import StandfordNLPProcessor
@@ -25,7 +26,7 @@ from ft.onto.base_ontology import Token, Sentence, Dependency
 
 
 def stanford_nlp_example(lang: str, text: str):
-    pl = Pipeline()
+    pl = Pipeline[DataPack]()
     pl.set_reader(StringReader())
 
     models_path = os.getcwd()
