+++ forte/processors/attribute_masking_processor.py
@@ -30,8 +30,8 @@ class AttributeMasker(PackProcessor):
     def initialize(self, _: Resources, config: HParams):
         self.fields = config.kwargs
 
-    @staticmethod
-    def default_configs() -> Dict[str, Any]:
+    @classmethod
+    def default_configs(cls) -> Dict[str, Any]:
         r"""Default config for this processor.
 
         Example usage is shown below
@@ -50,10 +50,12 @@ class AttributeMasker(PackProcessor):
             are the entry types whose fields need to be masked and the value is
             a list of field names.
         """
-        return {
+        config = super().default_configs()
+        config.update({
             "type": "",
             "kwargs": {}
-        }
+        })
+        return config
 
     def _process(self, input_pack: DataPack):
         for entry_type, attributes in self.fields:
