+++ forte/processors/base/batch_processor.py
@@ -16,7 +16,7 @@ The processors that process data in batch.
 """
 import itertools
 from abc import abstractmethod, ABC
-from typing import Dict, Optional, Type
+from typing import Dict, Optional, Type, Any
 
 from texar.torch import HParams
 
@@ -55,7 +55,6 @@ class BaseBatchProcessor(BaseProcessor[PackType], ABC):
         super().__init__()
         self.context_type: Type[Annotation] = self._define_context()
         self.input_info: DataRequest = self._define_input_info()
-
         self.batcher: ProcessingBatcher = self.define_batcher()
         self.use_coverage_index = False
 
@@ -65,13 +64,15 @@ class BaseBatchProcessor(BaseProcessor[PackType], ABC):
         assert configs is not None
         try:
             self.batcher.initialize(configs.batcher)
-        except AttributeError:
+        except AttributeError as e:
             raise ProcessorConfigError(
-                "To use batch processor, please provide the 'batch' "
-                "config at the top level.")
+                e, "Error in handling batcher config, please provide the "
+                   "check the config to see if you have the key 'batcher'."
+            )
 
+    @staticmethod
     @abstractmethod
-    def _define_context(self) -> Type[Annotation]:
+    def _define_context() -> Type[Annotation]:
         r"""User should define the context type for batch processors here. The
         context must be of type :class:`Annotation`, the processor will create
         data batches with in the span of each annotations. For example, if the
@@ -83,8 +84,9 @@ class BaseBatchProcessor(BaseProcessor[PackType], ABC):
         """
         raise NotImplementedError
 
+    @staticmethod
     @abstractmethod
-    def _define_input_info(self) -> DataRequest:
+    def _define_input_info() -> DataRequest:
         r"""User should define the :attr:`input_info` for the batch processors
         here. The input info will be used to get batched data for this
         processor.
@@ -94,8 +96,9 @@ class BaseBatchProcessor(BaseProcessor[PackType], ABC):
         """
         raise NotImplementedError
 
+    @staticmethod
     @abstractmethod
-    def define_batcher(self) -> ProcessingBatcher:
+    def define_batcher() -> ProcessingBatcher:
         r"""Define a specific batcher for this processor.
         Single pack :class:`BatchProcessor` initialize the batcher to be a
         :class:`~forte.data.batchers.ProcessingBatcher`.
@@ -176,6 +179,14 @@ class BaseBatchProcessor(BaseProcessor[PackType], ABC):
             self.pack(self.batcher.data_pack_pool[i], output_dict_i)
             start += self.batcher.current_batch_sources[i]
 
+    @classmethod
+    def default_configs(cls) -> Dict[str, Any]:
+        super_config = super().default_configs()
+
+        super_config['batcher'] = cls.define_batcher().default_configs()
+
+        return super_config
+
     @abstractmethod
     def pack(self, pack: PackType, inputs) -> None:
         r"""The function that task processors should implement.
@@ -233,8 +244,8 @@ class BatchProcessor(BaseBatchProcessor[DataPack], ABC):
 
 
 class FixedSizeBatchProcessor(BatchProcessor, ABC):
-
-    def define_batcher(self) -> ProcessingBatcher:
+    @staticmethod
+    def define_batcher() -> ProcessingBatcher:
         return FixedSizeDataPackBatcher()
 
 
@@ -257,6 +268,6 @@ class MultiPackBatchProcessor(BaseBatchProcessor[MultiPack], ABC):
 
 
 class FixedSizeMultiPackBatchProcessor(MultiPackBatchProcessor, ABC):
-
-    def define_batcher(self) -> ProcessingBatcher:
+    @staticmethod
+    def define_batcher() -> ProcessingBatcher:
         return FixedSizeDataPackBatcher()
