+++ tests/forte/data/readers/conll03_reader_test.py
@@ -36,7 +36,7 @@ class CoNLL03ReaderPipelineTest(unittest.TestCase):
         # Define and config the Pipeline
         self.dataset_path = "data_samples/conll03"
 
-        self.nlp = Pipeline()
+        self.nlp = Pipeline[DataPack]()
 
         self.nlp.set_reader(CoNLL03Reader())
         self.nlp.add_processor(DummyPackProcessor())
