+++ tests/forte/data/readers/html_reader_test.py
@@ -35,11 +35,11 @@ class HTMLReaderPipelineTest(unittest.TestCase):
         self.reader = HTMLReader(cache_directory=self._cache_directory,
                                  append_to_cache=True)
 
-        self.pl1 = Pipeline()
+        self.pl1 = Pipeline[DataPack]()
         self.pl1.set_reader(self.reader)
         self.pl1.initialize()
 
-        self.pl2 = Pipeline()
+        self.pl2 = Pipeline[DataPack]()
         self.pl2.set_reader(HTMLReader(from_cache=True,
                                        cache_directory=self._cache_directory))
         self.pl2.initialize()
