+++ tests/forte/data/readers/ontonotes_reader_test.py
@@ -36,7 +36,7 @@ class OntonotesReaderPipelineTest(unittest.TestCase):
         # Define and config the Pipeline
         self.dataset_path = "data_samples/ontonotes/00"
 
-        self.nlp = Pipeline()
+        self.nlp = Pipeline[DataPack]()
 
         self.nlp.set_reader(OntonotesReader())
         self.nlp.add_processor(DummyPackProcessor())
