+++ tests/forte/data/readers/string_reader_test.py
@@ -19,6 +19,7 @@ import os
 import unittest
 from pathlib import Path
 
+from forte.data.data_pack import DataPack
 from forte.data.readers import StringReader
 from forte.pipeline import Pipeline
 
@@ -29,11 +30,11 @@ class StringReaderPipelineTest(unittest.TestCase):
         # Define and config the Pipeline
         self.dataset_path = "forte/data/readers/tests/"
 
-        self.pl1 = Pipeline()
+        self.pl1 = Pipeline[DataPack]()
         self._cache_directory = Path(os.path.join(os.getcwd(), "cache_data"))
         self.pl1.set_reader(StringReader())
 
-        self.pl2 = Pipeline()
+        self.pl2 = Pipeline[DataPack]()
         self.pl2.set_reader(StringReader())
 
         self.text = (
