+++ tests/forte/processors/attribute_masking_processor_test.py
@@ -4,6 +4,7 @@ import os
 
 from texar.torch import HParams
 
+from forte.data.data_pack import DataPack
 from forte.pipeline import Pipeline
 from forte.data.readers import CoNLL03Reader
 from forte.processors import AttributeMasker
@@ -13,7 +14,7 @@ from ft.onto.base_ontology import Token
 class TestAttributeMaskingProcessor(unittest.TestCase):
 
     def test_without_attribute_masker(self):
-        pl = Pipeline()
+        pl = Pipeline[DataPack]()
         pl.set_reader(CoNLL03Reader())
         pl.initialize()
 
@@ -23,7 +24,7 @@ class TestAttributeMaskingProcessor(unittest.TestCase):
                 self.assertIsNotNone(entry.ner)
 
     def test_attribute_masker(self):
-        pl = Pipeline()
+        pl = Pipeline[DataPack]()
         pl.set_reader(CoNLL03Reader())
         config = {
             "kwargs": {
