+++ tests/forte/processors/lowercaser_processor_test.py
@@ -16,6 +16,7 @@ Unit tests for LowerCaser processor.
 """
 import unittest
 
+from forte.data.data_pack import DataPack
 from forte.pipeline import Pipeline
 from forte.data.readers import StringReader
 from forte.processors.lowercaser_processor import LowerCaserProcessor
@@ -23,7 +24,7 @@ from forte.processors.lowercaser_processor import LowerCaserProcessor
 
 class TestLowerCaserProcessor(unittest.TestCase):
     def setUp(self):
-        self.nlp = Pipeline()
+        self.nlp = Pipeline[DataPack]()
         self.nlp.set_reader(StringReader())
         self.nlp.add_processor(LowerCaserProcessor())
         self.nlp.initialize()
