+++ tests/forte/processors/stanfordnlp_processor_test.py
@@ -17,6 +17,7 @@ Unit tests for Stanford NLP processors.
 import os
 import unittest
 
+from forte.data.data_pack import DataPack
 from forte.pipeline import Pipeline
 from forte.data.readers import StringReader
 from forte.processors.stanfordnlp_processor import StandfordNLPProcessor
@@ -24,7 +25,7 @@ from forte.processors.stanfordnlp_processor import StandfordNLPProcessor
 
 class TestStanfordNLPProcessor(unittest.TestCase):
     def setUp(self):
-        self.stanford_nlp = Pipeline()
+        self.stanford_nlp = Pipeline[DataPack]()
         self.stanford_nlp.set_reader(StringReader())
         models_path = os.getcwd()
         config = {
